refactor(pkg/prp): update by cr
